use new schema for registry
